@@ -18,16 +18,16 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   # https://github.com/rrousselGit/river_pod
-  flutter_riverpod: ^0.12.4
+  flutter_riverpod: ^0.13.1
   # hook
-  flutter_hooks: ^0.15.0
-  hooks_riverpod: ^0.12.4
+  flutter_hooks: ^0.16.0
+  hooks_riverpod: ^0.13.1
 
 dev_dependencies:
   flutter_test:
