@@ -31,7 +31,7 @@ class ScopeProviderExample extends StatelessWidget {
 
 class ProductItem extends ConsumerWidget {
 
-  const ProductItem({Key key}): super(key: key);
+  const ProductItem({Key? key}): super(key: key);
 
   @override
   Widget build(BuildContext context, ScopedReader watch) {
